Refactored NavBar
