+++ src/components/Header/NavComponents/links.tsx
@@ -1,0 +1,24 @@
+import React from "react";
+import { NavTabLink } from "./tab-link-style";
+
+export const Links: React.FC<any> = props => {
+  return (
+    <React.Fragment>
+      <NavTabLink exact to="/dashboard">
+        DASHBOARD
+      </NavTabLink>
+      <NavTabLink exact to="/markets">
+        MARKETS
+      </NavTabLink>
+      <NavTabLink exact to="/borrow">
+        BORROW
+      </NavTabLink>
+      <NavTabLink exact to="/deposit">
+        DEPOSIT
+      </NavTabLink>
+      <NavTabLink exact to="/stake">
+        STAKE
+      </NavTabLink>
+    </React.Fragment>
+  );
+};
